refactor(mmdet): yaml alias and multi-dev
We have two possible sizes of the dev env. We added the possibility to
have the CUDA/CUDNN lib in the dev env now. We needed for profiling.
Also we explicitly set service for each target in the multi-build
Dockerfile. It is just more readable (and verbose) it does change much.

Eventually, we added tenorboard plugin as part of the dev dependencies.